requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
futu-api==8.7.4708
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.0.2
pandas==2.2.3
protobuf==3.20.3
psycopg2-binary==2.9.9
pycryptodome==3.21.0
python-dateutil==2.9.0.post0
pytz==2024.2
simplejson==3.19.3
six==1.16.0
tzdata==2024.2
Werkzeug==3.1.3
zipp==3.20.2
